Include only tests for supported architectures
