+++ tests/test_skipdata.c
@@ -6,86 +6,81 @@
 #include "../myinttypes.h"
 
 #include <capstone.h>
+#include "test_infra.h"
 
-struct platform {
-	cs_arch arch;
-	cs_mode mode;
-	unsigned char *code;
-	size_t size;
-	char *comment;
-	cs_opt_type opt_type;
-	cs_opt_value opt_value;
-	cs_opt_type opt_skipdata;
-	size_t skipdata;
-};
 
-static void print_string_hex(unsigned char *str, size_t len)
-{
-	unsigned char *c;
-
-	printf("Code: ");
-	for (c = str; c < str + len; c++) {
-		printf("0x%02x ", *c & 0xff);
-	}
-	printf("\n");
-}
-
+#ifdef CAPSTONE_ARM_SUPPORT
 static size_t mycallback(const uint8_t *buffer, size_t buffer_size, size_t offset, void *p)
 {
-	// always skip 2 bytes when encountering data
-	return 2;
+    // always skip 2 bytes when encountering data
+    return 2;
 }
+#endif
 
 static void test()
 {
+#ifdef CAPSTONE_X86_SUPPORT
 #define X86_CODE32 "\x8d\x4c\x32\x08\x01\xd8\x81\xc6\x34\x12\x00\x00\x00\x91\x92"
+#endif
 #define RANDOM_CODE "\xed\x00\x00\x00\x00\x1a\x5a\x0f\x1f\xff\xc2\x09\x80\x00\x00\x00\x07\xf7\xeb\x2a\xff\xff\x7f\x57\xe3\x01\xff\xff\x7f\x57\xeb\x00\xf0\x00\x00\x24\xb2\x4f\x00\x78"
 
-	cs_opt_skipdata skipdata = {
-		// rename default "data" instruction from ".byte" to "db"
-		"db",
-	};
+#if defined(CAPSTONE_X86_SUPPORT) || defined(CAPSTONE_ARM_SUPPORT)
+    cs_opt_skipdata skipdata = {
+        // rename default "data" instruction from ".byte" to "db"
+        "db",
+    };
+#endif
 
-	cs_opt_skipdata skipdata_callback = {
-		"db",
-		&mycallback,
-	};
+#ifdef CAPSTONE_ARM_SUPPORT
+    cs_opt_skipdata skipdata_callback = {
+        "db",
+        &mycallback,
+    };
+#endif
 
 	struct platform platforms[] = {
-		{
-			CS_ARCH_X86,
-			CS_MODE_32,
-			(unsigned char*)X86_CODE32,
-			sizeof(X86_CODE32) - 1,
-			"X86 32 (Intel syntax) - Skip data",
-		},
-		{
-			CS_ARCH_ARM,
-			CS_MODE_ARM,
-			(unsigned char*)RANDOM_CODE,
-			sizeof(RANDOM_CODE) - 1,
-			"Arm - Skip data",
-		},
-		{
-			CS_ARCH_X86,
-			CS_MODE_32,
-			(unsigned char*)X86_CODE32,
-			sizeof(X86_CODE32) - 1,
-			"X86 32 (Intel syntax) - Skip data with custom mnemonic",
-			0, 0,
-			CS_OPT_SKIPDATA_SETUP,
-			(size_t) &skipdata,
-		},
-		{
-			CS_ARCH_ARM,
-			CS_MODE_ARM,
-			(unsigned char*)RANDOM_CODE,
-			sizeof(RANDOM_CODE) - 1,
-			"Arm - Skip data with callback",
-			0, 0,
-			CS_OPT_SKIPDATA_SETUP,
-			(size_t) &skipdata_callback,
-		},
+#ifdef CAPSTONE_X86_SUPPORT
+    {
+        CS_ARCH_X86,
+        CS_MODE_32,
+        (unsigned char*)X86_CODE32,
+        sizeof(X86_CODE32) - 1,
+        "X86 32 (Intel syntax) - Skip data",
+    },
+#endif
+#ifdef CAPSTONE_ARM_SUPPORT
+    {
+        CS_ARCH_ARM,
+        CS_MODE_ARM,
+        (unsigned char*)RANDOM_CODE,
+        sizeof(RANDOM_CODE) - 1,
+        "Arm - Skip data",
+    },
+#endif
+#ifdef CAPSTONE_X86_SUPPORT
+    {
+        CS_ARCH_X86,
+        CS_MODE_32,
+        (unsigned char*)X86_CODE32,
+        sizeof(X86_CODE32) - 1,
+        "X86 32 (Intel syntax) - Skip data with custom mnemonic",
+        0, 0,
+        CS_OPT_SKIPDATA_SETUP,
+        (size_t) &skipdata,
+    },
+#endif
+#ifdef CAPSTONE_ARM_SUPPORT
+    {
+        CS_ARCH_ARM,
+        CS_MODE_ARM,
+        (unsigned char*)RANDOM_CODE,
+        sizeof(RANDOM_CODE) - 1,
+        "Arm - Skip data with callback",
+        0, 0,
+        CS_OPT_SKIPDATA_SETUP,
+        (size_t) &skipdata_callback,
+    },
+#endif
 	};
 
 	csh handle;
